requirements.txt
openai
discord.py
boto3
Pillow